requirements-lock.txt
accelerate==0.32.1
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiolimiter==1.1.0
aiosignal==1.3.1
airportsdata==20241001
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
astor==0.8.1
astroid==3.2.2
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.4.2
blake3==1.0.2
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==3.0.0
cmake==3.30.0
colorful==0.5.6
comm==0.2.2
compressed-tensors==0.8.1
confection==0.1.5
contourpy==1.2.1
coverage==7.5.4
cycler==0.12.1
cymem==2.0.8
datasets==2.20.0
debugpy==1.8.2
decorator==5.1.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.6.1
einops==0.8.0
email_validator==2.2.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
et-xmlfile==1.1.0
eval_type_backport==0.2.0
evaluate==0.4.2
executing==2.0.1
fastapi==0.114.2
fastapi-cli==0.0.4
fastchat==0.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
gdown==5.2.0
gguf==0.10.0
google-api-core==2.24.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
h11==0.14.0
hdbscan==0.8.38.post1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.27.1
idna==3.7
importlib_metadata==8.0.0
-e git+ssh://git@github.com/jantrienes/info-salience.git@6e6c9f9126692ba457530ed7487c6a8a3aa4233c#egg=info_salience
iniconfig==2.0.0
inquirerpy==0.3.4
InstructorEmbedding @ git+ssh://git@github.com/pascalhuerten/instructor-embedding.git@4721e7375afeb8fcb32400a13057f9348bb69392
interegular==0.3.3
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json_repair==0.30.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
krippendorff==0.7.0
label-studio-sdk==0.0.34
label-studio-tools==0.0.4
langcodes==3.4.0
language_data==1.2.0
lark==1.2.2
linkify-it-py==2.0.3
litellm==1.49.6
llvmlite==0.43.0
lm-format-enforcer==0.10.9
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
minicheck @ git+ssh://git@github.com/jantrienes/MiniCheck.git@451a3dab7a192eeeed9a87d65a392f1f5dd9880a
mistral_common==1.5.1
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.18.6
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
nbformat==5.10.4
nervaluate==0.2.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.59.9
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.6
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
pandas==2.2.2
parso==0.8.4
partial-json-parser==0.2.1.1.post4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.0
pluggy==1.5.0
preshed==3.0.9
prometheus-eval==0.1.20
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pylint==3.2.5
pynndescent==0.5.13
pyparsing==3.1.2
pyphen==0.15.0
pysbd==0.3.4
PySocks==1.7.1
pytest==8.2.2
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
ray==2.31.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
summa==1.2.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
textstat==0.7.3
textual==1.0.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
together==1.2.1
tokenize-rt==5.2.0
tokenizers==0.21.0
tomlkit==0.12.5
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.48.1
triton==3.1.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
umap-learn==0.5.6
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
virtualenv==20.29.1
vllm==0.6.6.post1
wasabi==1.1.3
watchfiles==0.22.0
wcwidth==0.2.13
weasel==0.4.1
websockets==12.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xformers==0.0.28.post3
xgrammar==0.1.10
xmljson==0.2.1
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
